Update configurator.py
Fix for:SPP-194
The response time for "api/config" call varies mainly because of getting fluentd version . The command td-agent --version takes anywhere between 1-15s to response.

The version is not used in any place. Removed the version details from the api response.

Testcase:
Tested configurator is u and running and response of the api is less than or equal to 1 s.